Restructure Flax train functionality
