+++ scico/test/flax/test_inv.py
@@ -11,7 +11,7 @@ import pytest
 from scico import flax as sflax
 from scico import random
 from scico.flax.examples import PaddedCircularConvolve, build_blur_kernel, have_astra
-from scico.flax.train.train import clip_positive, clip_range, construct_traversal
+from scico.flax.train.traversals import clip_positive, clip_range, construct_traversal
 from scico.linop import CircularConvolve, Identity
 
 if have_astra:
